backend/requirements_template_editor.txt
# Template Editor Agent Dependencies
# Additional requirements for the template editor designer agent

# LangGraph for workflow management
langgraph>=0.0.20

# Image processing
Pillow>=10.0.0

# OpenAI for vision and content generation
openai>=1.0.0

# Additional image manipulation
opencv-python>=4.8.0

# Template processing
jinja2>=3.1.0

# Image generation and manipulation
diffusers>=0.21.0
transformers>=4.30.0

# Advanced image processing
scikit-image>=0.21.0

# Color manipulation
colorthief>=0.2.1

# Font handling
fonttools>=4.40.0

# Template matching
opencv-contrib-python>=4.8.0

# Additional utilities
python-multipart>=0.0.6
aiofiles>=23.0.0
